Unified error-behavior for RecordStore implementations Added RecordStore#getAllValues Added extra exception-types for missing table/attribute Added Mixin to test Exceptions Signed-off-by:doe300 <stadeldani@web.de>

+++ test/de/doe300/activerecord/record/association/CachedRecordSetTest.java
@@ -191,7 +191,7 @@ public class CachedRecordSetTest extends Assert
 	public void testAggregate()
 	{
 		final CountNotNull<TestInterface, String> count = new CountNotNull<>("name", TestInterface::getName);
-		assertEquals( base.aggregate( count, null ), base.getAll().cached().aggregate( count, null));
+		assertEquals( base.aggregate( count, null ).longValue(), base.getAll().cached().aggregate( count, null).longValue());
 	}
 
 	@Test
